Move Logic and LogicManager to commons.logic
